Moving around a few items within \rtos
