+++ rtos/include/rtos/EventFlags.h
@@ -26,9 +26,9 @@
 #include <cstddef>
 #include <stdint.h>
 #include "rtos/Kernel.h"
-#include "rtos/mbed_rtos_types.h"
-#include "rtos/mbed_rtos1_types.h"
-#include "rtos/mbed_rtos_storage.h"
+#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/internal/mbed_rtos1_types.h"
+#include "rtos/internal/mbed_rtos_storage.h"
 
 #include "platform/NonCopyable.h"
