+++ rtos/include/rtos/Mail.h
@@ -28,9 +28,9 @@
 
 #include "rtos/Queue.h"
 #include "rtos/MemoryPool.h"
-#include "rtos/mbed_rtos_types.h"
-#include "rtos/mbed_rtos_storage.h"
-#include "rtos/mbed_rtos1_types.h"
+#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/internal/mbed_rtos_storage.h"
+#include "rtos/internal/mbed_rtos1_types.h"
 
 #include "platform/mbed_toolchain.h"
 #include "platform/mbed_assert.h"
