+++ rtos/include/rtos/Mutex.h
@@ -24,9 +24,9 @@
 #define MUTEX_H
 
 #include <stdint.h>
-#include "rtos/mbed_rtos_types.h"
-#include "rtos/mbed_rtos1_types.h"
-#include "rtos/mbed_rtos_storage.h"
+#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/internal/mbed_rtos1_types.h"
+#include "rtos/internal/mbed_rtos_storage.h"
 #include "rtos/Kernel.h"
 
 #include "platform/NonCopyable.h"
