+++ rtos/include/rtos/Semaphore.h
@@ -25,9 +25,9 @@
 
 #include <stdint.h>
 #include <chrono>
-#include "rtos/mbed_rtos_types.h"
-#include "rtos/mbed_rtos1_types.h"
-#include "rtos/mbed_rtos_storage.h"
+#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/internal/mbed_rtos1_types.h"
+#include "rtos/internal/mbed_rtos_storage.h"
 #include "rtos/Kernel.h"
 #include "platform/mbed_toolchain.h"
 #include "platform/NonCopyable.h"
