+++ rtos/source/rtos_handlers.h
@@ -23,7 +23,7 @@
 #ifndef RTOS_HANDLERS_H
 #define RTOS_HANDLERS_H
 
-#include "rtos/mbed_rtos_types.h"
+#include "rtos/internal/mbed_rtos_types.h"
 
 #ifdef __cplusplus
 extern "C" {
