+++ rtos/tests/TESTS/mbedmicro-rtos-mbed/event_flags/main.cpp
